split out helpers from main method file
